mappings object momdel <=> sql statmenets

+++ javakurs3-bibliothek-jsf-mobile/src/main/java/at/bibliothek/services/database/RentDBService.java
@@ -90,7 +90,7 @@ public class RentDBService {
 	 * @param custId
 	 * @return
 	 */
-	public int getRentsCount( Map<String, Object> filters, Long custId) {
+	public long getRentsCount( Map<String, Object> filters, Long custId) {
 
 		if (custId !=null){
 			filters.put("kunde.id", custId);
@@ -99,7 +99,7 @@ public class RentDBService {
 		String queryString =  " select count (*) from Rent r " + composeWhereClause(filters);
 
 		Query loadQuery = em.createQuery( queryString );
-		List<Integer> result  = loadQuery.getResultList();
+		List<Long> result  = loadQuery.getResultList();
 		
 		return result.size() > 0 ? result.get(0) : 0;
 	}
